requirements.txt
torch
torchvision
numpy
nltk
streamlit
Pillow
tqdm
